yak shave: include py.typed and move to src layout
In PEP 561, it is stated that we need to include the file py.typed in
order for type checkers to be able to use the type hints:

https://www.python.org/dev/peps/pep-0561/#packaging-type-information

It also says that module-only distributions are not supported and so
the code needs to be refactored into a package-based distribution.
Therefore I moved jdcal.py into its own package directory. This then
necessitated importing its functions in the __init__.py so that the API
won't change (i.e. keep jdcal.is_leap rather than jdcal.jdcal.is_leap).

But mypy --strict isn't happy with this unless you also explicitly re-
export everything via __all__ or do from X import Y as Y on everything.
I don't understand why this extra layer of redundancy is required but
I can't find a way around it:

https://mypy.readthedocs.io/en/latest/config_file.html#confval-implicit
_reexport